docs/requirements.txt
# Necessary
sphinx
qiskit-sphinx-theme
reno
pandas
matplotlib
nbsphinx
sphinx_reredirects
sphinx-remove-toctrees
sphinx-copybutton
numpydoc
sphinx-design
sphinx-autodoc-typehints
sphinx-togglebutton
sphinxcontrib-svg2pdfconverter
docutils
pygments-pytest
anybadge
spinx-toolbox
sphinxcontrib-plot
